refactor(discovery): promote internal/discovery to public pkg/discovery
internal/discovery (mDNS browse/advertise: Manager, Config, ServerInfo,
ClientInfo, NewManager) is the maintained discovery implementation, used by
both the server (advertise + browse-for-clients) and the player CLI
(browse-for-servers). For the repo split the CLI repos must reach it across
a module boundary, where Go forbids importing another module's internal/.

Move it verbatim to pkg/discovery (the name Phase 0 freed by deleting the
orphaned stale fork) and repoint the four importers (main.go, server.go,
client_dialer.go + test). No behavior change; pkg/sendspin no longer imports
internal/discovery.

Phase 2 of the three-repo split.
