+++ pkg/discovery/txt.go
