+++ pkg/discovery/txt_test.go
