+++ pkg/sendspin/client_dialer.go
@@ -12,7 +12,7 @@ import (
 	"sync"
 	"time"
 
-	"github.com/Sendspin/sendspin-go/internal/discovery"
+	"github.com/Sendspin/sendspin-go/pkg/discovery"
 	"github.com/gorilla/websocket"
 )
